% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BPL.R
\name{BPL}
\alias{BPL}
\alias{dBPL}
\alias{pBPL}
\alias{qBPL}
\alias{rBPL}
\title{Individual Size Distribution for fitting a GAMLSS}
\usage{
BPL(counts, lower, upper, mu.link = "identity")

dBPL(x = 1, mu = -2, lower = 1, upper = 100, log = FALSE)

pBPL(q = 1, mu = -2, lower = 1, upper = 100, log = FALSE)

qBPL(p = 0.1, mu = -2, lower = 1, upper = 100)

rBPL(n = 1, mu = -2, lower = 1, upper = 100)
}
\arguments{
\item{counts}{Number of observations for a corresponding size}

\item{lower}{Lowest size limit (does not need to equal `min(size)`)}

\item{upper}{Highest size limit (does not need to equal `max(size)`)}

\item{mu.link}{The exponent of ISD}
}
\value{
A `gamlss.family` object
}
\description{
`BPL()` defines the bounded Power-Law distribution within the `gamlss` environment (as the equivalent of `bPL()` in `gamlss2`)
}
\examples{

d = data.frame(size = rbPL(1e4))

# estimate the exponent 'b'
# using 'gamlss2'
m2 = gamlss2(size ~ 1,
             data = d,
             family = bPL(counts = 1, lower = 1, upper = 1000))
summary(m2)
# using 'gamlss'
m = gamlss(size ~ 1,
           data = d,
           family = BPL(counts = 1, lower = 1, upper = 1000))
summary(m)

}
\references{
Edwards, A.M., Robinson, J.P.W., Plank, M.J., Baum, J.K. and Blanchard, J.L. (2017), Testing and recommending methods for fitting size spectra to data. Methods Ecol Evol, 8: 57-67. https://doi.org/10.1111/2041-210X.12641

Edwards, A. M., Robinson, J. P. W., Blanchard, J. L., Baum, J. K., & Plank, M. J. (2020). Accounting for the bin structure of data removes bias when fitting size spectra. Marine Ecology Progress Series, 636, 19–33. https://www.jstor.org/stable/26920653
}
