% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bPL_functions.R
\name{dbPL}
\alias{dbPL}
\title{bounded Power-Law}
\usage{
dbPL(x = 1, b = -2, lower = 1, upper = 100, log = FALSE)
}
\arguments{
\item{x}{Vector of values to compute density}

\item{b}{Exponent of ISD}

\item{lower}{Lowest size limit}

\item{upper}{Highest size limit}

\item{log}{Use `log`?}

\item{q}{Vector of values to compute distribution function}

\item{p}{Vector of probabilities}
}
\value{
dbPL returns pdf. pbPL returns cumulative distribution values. qbPL gives the quantiles. rbPL generates random numbers.
}
\description{
Density, Distribution function, quantile function and, random generation.
Mostly based on `sizeSpectra` package; they have been re-written here to capture all essentials for `gamlss2` functionality and exported for convenience.
}
\examples{

# none

}
\references{
Edwards, A.M., Robinson, J.P.W., Plank, M.J., Baum, J.K. and Blanchard, J.L. (2017), Testing and recommending methods for fitting size spectra to data. Methods Ecol Evol, 8: 57-67. https://doi.org/10.1111/2041-210X.12641

Edwards, A. M., Robinson, J. P. W., Blanchard, J. L., Baum, J. K., & Plank, M. J. (2020). Accounting for the bin structure of data removes bias when fitting size spectra. Marine Ecology Progress Series, 636, 19–33. https://www.jstor.org/stable/26920653
}
