% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounded_power_law.R
\name{bPL}
\alias{bPL}
\title{Individual Size Distribution for fitting a GAMLSS}
\usage{
bPL(counts, lower, upper)
}
\arguments{
\item{counts}{Number of observations for a corresponding size}

\item{lower}{Lowest size limit (does not need to equal `min(size)`)}

\item{upper}{Highest size limit (does not need to equal `max(size)`)}
}
\value{
A `gamlss2.family` object
}
\description{
`bPL` returns a distribution function for ISD described as a bounded Power-Law distribution (bPL) in
Edwards et al. 2017. `bPL` is defined by only one parameter (defined as `mu` here) which is equivalent to 'b'
as described in Edwards et al. 2017. When fitting with `gamlss2`, estimate for `mu` is returned.
}
\examples{

# simulate a dataset of 1000 sizes coming from an ISD with
# b = -2, and range of sizes bounded between 1 and 1000

# 1) either use sizeSpectra::rPLB
d = data.frame(size = sizeSpectra::rPLB(1e4, -2, 1, 1000))
# 2) or, alternatively, use `bPL()$r`
d = data.frame(size = bPL(1, 1, 1000)$r(1e4, list(mu = -2)))

# estimate the exponent 'b'
m = gamlss2(size ~ 1, data = d, family = bPL(counts = 1, lower = 1, upper = 1000))

summary(m)

}
\references{
Edwards, A.M., Robinson, J.P.W., Plank, M.J., Baum, J.K. and Blanchard, J.L. (2017), Testing and recommending methods for fitting size spectra to data. Methods Ecol Evol, 8: 57-67. https://doi.org/10.1111/2041-210X.12641

Edwards, A. M., Robinson, J. P. W., Blanchard, J. L., Baum, J. K., & Plank, M. J. (2020). Accounting for the bin structure of data removes bias when fitting size spectra. Marine Ecology Progress Series, 636, 19–33. https://www.jstor.org/stable/26920653
}
